anvesha-otel/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anvesha-otel"
version = "0.1.0"
description = "OpenTelemetry registration helpers for Anvesha tracing."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openinference-semantic-conventions>=0.1.12",
  "opentelemetry-api>=1.29.0",
  "opentelemetry-sdk>=1.29.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.29.0",
  "opentelemetry-exporter-otlp-proto-http>=1.29.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/anvesha"]
